requirements.txt
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2==2.7.7
python-dotenv==0.10.1
SQLAlchemy==1.2.17
Werkzeug==0.14.1
